+++ src/sub_account/rest_api/apis/asset_management_api.rs
@@ -435,6 +435,11 @@ pub struct GetSubAccountDepositHistoryParams {
     /// This field is **required.
     #[builder(setter(into))]
     pub email: String,
+    /// Default: `false`, return `sourceAddress`field when set to `true`
+    ///
+    /// This field is **optional.
+    #[builder(setter(into), default)]
+    pub include_source: Option<bool>,
     ///
     /// The `coin` parameter.
     ///
@@ -1564,6 +1569,7 @@ impl AssetManagementApi for AssetManagementApiClient {
     {
         let GetSubAccountDepositHistoryParams {
             email,
+            include_source,
             coin,
             status,
             start_time,
@@ -1578,6 +1584,10 @@ impl AssetManagementApi for AssetManagementApiClient {
         let body_params = BTreeMap::new();
 
         query_params.insert("email".to_string(), json!(email));
+
+        if let Some(rw) = include_source {
+            query_params.insert("includeSource".to_string(), json!(rw));
+        }
 
         if let Some(rw) = coin {
             query_params.insert("coin".to_string(), json!(rw));
@@ -3441,7 +3451,7 @@ mod tests {
         TOKIO_SHARED_RT.block_on(async {
             let client = MockAssetManagementApiClient { force_error: false };
 
-            let params = GetSubAccountDepositHistoryParams::builder("sub-account-email@email.com".to_string(),).coin("coin_example".to_string()).status(789).start_time(1623319461670).end_time(1641782889000).limit(1).offset(0).recv_window(5000).tx_id("1".to_string()).build().unwrap();
+            let params = GetSubAccountDepositHistoryParams::builder("sub-account-email@email.com".to_string(),).include_source(false).coin("coin_example".to_string()).status(789).start_time(1623319461670).end_time(1641782889000).limit(1).offset(0).recv_window(5000).tx_id("1".to_string()).build().unwrap();
 
             let resp_json: Value = serde_json::from_str(r#"[{"id":"769800519366885376","amount":"0.001","coin":"BNB","network":"BNB","status":0,"address":"bnb136ns6lfw4zs5hg4n85vdthaad7hq5m4gtkgf23","addressTag":"101764890","txId":"98A3EA560C6B3336D348B6C83F0F95ECE4F1F5919E94BD006E5BF3BF264FACFC","insertTime":1661493146000,"transferType":0,"confirmTimes":"1/1","unlockConfirm":0,"walletType":0},{"id":"769754833590042625","amount":"0.50000000","coin":"IOTA","network":"IOTA","status":1,"address":"SIZ9VLMHWATXKV99LH99CIGFJFUMLEHGWVZVNNZXRJJVWBPHYWPPBOSDORZ9EQSHCZAMPVAPGFYQAUUV9DROOXJLNW","addressTag":"","txId":"ESBFVQUTPIWQNJSPXFNHNYHSQNTGKRVKPRABQWTAXCDWOAKDKYWPTVG9BGXNVNKTLEJGESAVXIKIZ9999","insertTime":1599620082000,"transferType":0,"confirmTimes":"1/1","unlockConfirm":0,"walletType":0}]"#).unwrap();
             let expected_response : Vec<models::GetSubAccountDepositHistoryResponseInner> = serde_json::from_value(resp_json.clone()).expect("should parse into Vec<models::GetSubAccountDepositHistoryResponseInner>");